--- server/routes/dbRouter.js
@@ -1,19 +1,0 @@
-const express = require('express');
-
-const dbController = require('../controllers/dbController');
-
-const router = express.Router();
-
-router.get('/', dbController.getTest, (req, res) => {
-  return res.status(200).send(res.locals);
-});
-
-// router.post('/', dbController.saveAlgo, (req, res) => {
-//   return res.status(200).send(res.locals);
-// });
-
-// router.delete('/', dbController.delAlgo, (req, res) => {
-//   return res.status(200).send(res.locals);
-// });
-
-module.exports = router;
