+++ server/server.js
@@ -4,7 +4,7 @@ const express = require('express');
 const app = express();
 const path = require('path');
 
-const dbRouter = require('./routes/dbRouter');
+const userRouter = require('./routes/userRouter');
 
 app.use(express.json());
 
@@ -16,6 +16,6 @@ app.get('/', (req, res) => {
   return res.status(200).sendFile(path.join(__dirname, '..', 'client', 'index.html'));
 });
 
-app.use('/api', dbRouter);
+app.use('/api', userRouter);
 
 app.listen(3000);
